Metamorph/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch==2.1.0
torch-geometric==2.4.0
geopandas==0.14.1
shapely==2.0.2
rasterio==1.3.9
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
folium==0.15.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
networkx==3.2.1
osmnx==1.7.1
gunicorn==21.2.0